requirements.txt
Flask
Flask-Login
requests
python-dotenv
Flask-Caching
gunicorn
SQLAlchemy
psycopg[binary]
